+++ Changelog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented here.
 - Improve display of Python package installation errors when creating environment (#585)
 - Update "setting up test environment" message with http response of status code 503 (#589)
 - Change rlimit resource settings to apply each worker individually (#587) 
+- Improve error reporting with handled assertion errors (#591)
 
 ## [v2.6.0]
 - Update python versions in docker file (#568)